I forgot in previous commit: added new files. Solved a bug affecting currentLine for which it was impossible to get out of 256 line patterns. (I'm a genius btw)

+++ source/arm7/source/libxm7.h
@@ -177,7 +177,7 @@ typedef struct
 
     u8 CurrentSongPosition; //  the pattern in playback now (position in the PatternOrder array)
     u8 CurrentPatternNumber;       //  the pattern in playback now (its number!)
-    u8 CurrentLine;      //  the line in playback now (0..length of the pattern)
+    u16 CurrentLine;      //  the line in playback now (0..length of the pattern)
     u8 CurrentTick;                         //  the tick in playback now (0..31)
 
     u8 CurrentDelayLines;  //  the lines to "delay" after this line (effect EEx)
